requirements.txt
requests==2.26.0
python-telegram-bot==13.7
python_dotenv==0.19.0
